fix: getSections blocks in reducer file get seccions blocks to store in spec document
